--- src/a2a.rs
@@ -1,7 +1,0 @@
-use reqwest::Url;
-
-#[derive(Debug)]
-pub struct Client {
-    pub url: Url,
-    pub client: reqwest::Client,
-}
