+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod inbound;
 pub mod metrics;
 pub mod outbound;
 pub use outbound::backend;
+pub mod a2a;
 pub mod proto;
 pub mod proxyprotocol;
 pub mod rbac;
@@ -16,5 +17,3 @@ pub mod strng;
 pub mod trcng;
 pub mod version;
 pub mod xds;
-pub mod a2a;
-pub mod a2a_relay;
